scugic: Typecast with u32 to fix misrac violation
Typecast with u32 to match the lefthand and righthand operands
of the code and fix the misra-c violation.

misra_c_2012_rule_10_4_violation: Essential type of the left
hand operand (unsigned) is not the same as that of the right
operand (signed).


Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.paladugu@xilinx.com>

+++ XilinxProcessorIPLib/drivers/scugic/src/xscugic.c
@@ -1186,7 +1186,7 @@ void XScuGic_Stop(XScuGic *InstancePtr)
 
 		Target_Cpu = XScuGic_DistReadReg(InstancePtr,
 					XSCUGIC_SPI_TARGET_OFFSET_CALC(Int_Id));
-		if ((Target_Cpu != LocalCpuID) && (Target_Cpu != 0)) {
+		if ((Target_Cpu != LocalCpuID) && (Target_Cpu != (u32)0)) {
 			/*
 			 * If any other CPU is also programmed to target
 			 * register, GIC distributor can not be disabled.
@@ -1205,7 +1205,7 @@ void XScuGic_Stop(XScuGic *InstancePtr)
 	 * If GIC distributor is safe to be disabled, disable all the interrupt
 	 * and then disable distributor.
 	 */
-	if (DistDisable == 1) {
+	if (DistDisable == (u32)1) {
 		for (Int_Id = 0U; Int_Id < XSCUGIC_MAX_NUM_INTR_INPUTS;
 				Int_Id = Int_Id+32U) {
 			/*
